requirement.txt
langchain
langchain-core
langchain-community

langchain-pinecone
langchain-google-genai

youtube-transcript-api
pytubefix

langchain-google-genai
google-generativeai

python-dotenv